Move collectEnvironmentVariables and isValidEnvKey methods to digdag-plugin-utils
