+++ digdag-plugin-utils/src/main/java/io/digdag/util/CommandOperators.java
@@ -1,0 +1,27 @@
+package io.digdag.util;
+
+import io.digdag.client.config.ConfigException;
+import io.digdag.spi.PrivilegedVariables;
+import java.util.Map;
+import java.util.regex.Pattern;
+
+public class CommandOperators
+{
+    private final static Pattern VALID_ENV_KEY = Pattern.compile("[a-zA-Z_][a-zA-Z_0-9]*");
+
+    public static void collectEnvironmentVariables(final Map<String, String> env, final PrivilegedVariables variables)
+    {
+        for (String name : variables.getKeys()) {
+            if (!VALID_ENV_KEY.matcher(name).matches()) {
+                throw new ConfigException("Invalid _env key name: " + name);
+            }
+            env.put(name, variables.get(name));
+        }
+    }
+
+    public static boolean isValidEnvKey(String key)
+    {
+        return VALID_ENV_KEY.matcher(key).matches();
+    }
+
+}
