+++ digdag-standards/src/main/java/io/digdag/standards/operator/RbOperatorFactory.java
@@ -14,16 +14,15 @@ import io.digdag.client.config.ConfigElement;
 import io.digdag.spi.CommandExecutorContext;
 import io.digdag.spi.CommandExecutorRequest;
 import io.digdag.spi.CommandExecutor;
-import io.digdag.spi.CommandLogger;
 import io.digdag.spi.CommandStatus;
 import io.digdag.spi.OperatorContext;
 import io.digdag.spi.TaskExecutionException;
 import io.digdag.spi.TaskResult;
 import io.digdag.spi.Operator;
 import io.digdag.spi.OperatorFactory;
-import io.digdag.standards.command.SimpleCommandExecutor;
 import io.digdag.standards.operator.state.TaskState;
 import io.digdag.util.BaseOperator;
+import io.digdag.util.CommandOperators;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.IOException;
@@ -201,7 +200,7 @@ public class RbOperatorFactory
 
             final Map<String, String> environments = Maps.newHashMap();
             environments.putAll(System.getenv());
-            SimpleCommandExecutor.collectEnvironmentVariables(environments, context.getPrivilegedVariables());
+            CommandOperators.collectEnvironmentVariables(environments, context.getPrivilegedVariables());
 
             final CommandExecutorContext context = buildCommandExecutorContext(projectPath);
             final CommandExecutorRequest request = buildCommandExecutorRequest(context, workingDirectory, tempDir, environments, cmdline.build());
